encoder: calculate encoded video time in floating point

+++ source/encoder/encoder.cpp
@@ -352,7 +352,7 @@ void Encoder::fetchStats(x265_stats *stats)
         stats->globalPsnr = 0;
     }
     stats->elapsedEncodeTime = (double)(x265_mdate() - m_encodeStartTime) / 1000000;
-    stats->elapsedVideoTime = stats->encodedPictureCount / param.frameRate;
+    stats->elapsedVideoTime = (double)stats->encodedPictureCount / param.frameRate;
     stats->bitrate = (0.001f * stats->accBits) / stats->elapsedVideoTime;
     stats->totalWPFrames = m_numWPFrames;
 }
